Remove transform routes, some other stuff

+++ pyvrp/cpp/search/Route.h
@@ -34,9 +34,9 @@ class Route
     // Sets forward node time windows.
     void setupRouteTimeWindows();
 
-public:             // TODO make fields private
-    int const idx;  // Route index
-    Node *depot;    // Pointer to the associated depot
+public:                // TODO make fields private
+    size_t const idx;  // Route index
+    Node *depot;       // Pointer to the associated depot
 
     /**
      * @return The client or depot node at the given position.
